mergeable Update components.md
Adds a link to the code to compare the wrong way to, to learn the right way. It's also my opinion the typescript type annotations for props makes what's happening clear but also for novices can be too much information to keep in mind, so maybe a plain JS with props implementation of RevealAnswer would be better.